Move lobby search functionality to submodule
